refactor app
refactor main for app only and data reads
